load interface_driver using a symbolic name
Instead of relying on a full class path let's use symbolic names and
namespaces from setup.cfg which relies on stevedore to use a proper
class. This can only be used for code-paths that do not rely on direct
usage of importlib which is not the case for the metering agent, which
is why its config is left untouched.

Co-Authored-By: Anton Kremenetsky <akremenetsky@dev.rtsoft.ru>
Co-Authored-By: Dmitrii Shcherbakov <dmitrii.shcherbakov@canonical.com>
Change-Id: I4d3389a0fe376fed87265f51fdd69caf14fb3b16
Closes-Bug: #1747964